Polishing
- extract common abstract test specification for plugin func tests
- update javadocs
